chore: Fix cyclic deps issue
